+++ plugins/filetools/plugin.js
@@ -48,7 +48,9 @@
 			editor.on( 'fileUploadRequest', function( evt ) {
 				var fileLoader = evt.data.fileLoader,
 					$formData = new FormData(),
-					requestData = evt.data.requestData;
+					requestData = evt.data.requestData,
+					configXhrHeaders = editor.config.fileTools_requestHeaders,
+					header;
 
 				for ( var name in requestData ) {
 					var value = requestData[ name ];
@@ -63,6 +65,12 @@
 				}
 				// Append token preventing CSRF attacks.
 				$formData.append( 'ckCsrfToken', CKEDITOR.tools.getCsrfToken() );
+
+				if ( configXhrHeaders ) {
+					for ( header in configXhrHeaders ) {
+						fileLoader.xhr.setRequestHeader( header, configXhrHeaders[ header ] );
+					}
+				}
 
 				fileLoader.xhr.send( $formData );
 			}, null, null, 999 );
@@ -856,7 +864,22 @@
 		 */
 		isTypeSupported: function( file, supportedTypes ) {
 			return !!file.type.match( supportedTypes );
-		}
+		},
+
+		/**
+		 * Feature detection indicating whether current browser supports methods essential to send files over XHR request.
+		 *
+		 * @since 4.8.1
+		 * @property {Boolean} isFileUploadSupported
+		 */
+		isFileUploadSupported: ( function() {
+			return typeof FileReader === 'function' &&
+				typeof ( new FileReader() ).readAsDataURL === 'function' &&
+				typeof FormData === 'function' &&
+				typeof ( new FormData() ).append === 'function' &&
+				typeof XMLHttpRequest === 'function' &&
+				typeof Blob === 'function';
+		} )()
 	} );
 } )();
 
@@ -883,5 +906,22 @@
  *
  * @since 4.5.3
  * @cfg {String} [fileTools_defaultFileName='']
+ * @member CKEDITOR.config
+ */
+
+/**
+ * Allows to add extra headers for every request made using {@link CKEDITOR.fileTools} API.
+ *
+ * Note that headers can still be customized per a single request, using the
+ * [`fileUploadRequest`](https://docs.ckeditor.com/ckeditor4/docs/#!/api/CKEDITOR.editor-event-fileUploadRequest)
+ * event.
+ *
+ *		config.fileTools_requestHeaders = {
+ *			'X-Requested-With': 'XMLHttpRequest',
+ *			'Custom-Header': 'header value'
+ *		};
+ *
+ * @since 4.8.1
+ * @cfg {Object} [fileTools_requestHeaders]
  * @member CKEDITOR.config
  */